refactor(ci): extract CI functionality into dedicated crate with cargo alias
- Create new `ci` crate in its own directory with dedicated CI logic
- Add `cargo ci` alias in .cargo/config.toml for cleaner command interface
- Move all CI functionality from setup tool to dedicated ci tool
- Support `cargo ci check` and `cargo ci test` subcommands
- Update GitHub Actions workflow to use new `cargo ci` commands
- Clean up setup tool by removing CI-related code and subcommands
- Maintain separation of concerns: setup for development, ci for automation
