+++ .cargo/config.toml
@@ -1,2 +1,3 @@
 [alias]
 setup = "run -p setup --"
+ci = "run -p ci --"
