+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["symposium/mcp-server", "setup", "md-rfd-preprocessor"]
+members = ["symposium/mcp-server", "setup", "md-rfd-preprocessor", "ci"]
 resolver = "2"
 
 [workspace.dependencies]
